refactor: remove adapter
